+++ __test__/hooks/useVueRouter.spec.ts
@@ -1,0 +1,15 @@
+import setupMiniApp from '../utils/setupMiniApp'
+import { useVueRouter } from '../../src/hooks/web/useVueRouter'
+
+describe('useVueRouter', async () => {
+  await setupMiniApp()
+
+  const { router } = useVueRouter()
+
+  it('should get push and replace methods', () => {
+    const { push, replace } = router
+
+    assert.isFunction(push)
+    assert.isFunction(replace)
+  })
+})